services/orchestrator-agent/requirements.txt
fastapi[standard]
pydantic
langchain
langgraph
langchain-openai
langchain-community
httpx
motor
pymongo
aiokafka
python-dotenv
